==> 00-run-all.r <==
library(tidyverse)
library(lubridate)
library(zoo)
theme_set(theme_light())

source("01-config.r")
source("02-import-data.r")
source("03-listings-and-sales.r")
source("03-inventory.r")

source("03-sales-price.r")
source("03-time-on-market.r")
source("04-write-data.r")

source("05-find_resales.r")
source("06-dashboard.r")
source("05-predictions.r")
source("market-fundamentals.r")
source("more-predictions.r")
